Add continue, break support
Added stack to hold symbol categories (uses for symbols, e.g., using a
label to mark end of loop).
The stack is used so codegen for continue and return knows which label
to jump to by looking at the top (latest entry => closest loop) of the
stack. Push and pop entries on the stack on enter and exit loops.