Add computation of liveness use/def per block
An il instruction can define (assign to) symbols and it can use symbols,
use/def contains the symbols which are used in the block (thus required
from preceding blocks), and symbols defined (assigned) in the block
respectively.
Use/def information is stored in the block, computed by walking through
each statement in a block backwards, a symbol which was previously used
but now defined is removed from use (not required from other blocks).
Liveness will be used to build the interference graph for register
allocation.

- Add vec_splice function to remove elements from dynamic array