Add liveness analysis across blocks
IN are symbols which are needed coming into the block (used in
the block or subsequent blocks). OUT are the symbols which remain
needed upon leaving the block.
Because the flow graph is cyclic, the solution to determining symbols
needed in/out of blocks is to repeatedly traverse the control flow graph
and propagate liveness until the results are stable. Otherwise, give up
if it takes too many attempts, in which case assume the worst case for
liveness.
- Replace magic number 2 (maximum number of links for block) with macro
  MAX_BLOCK_LINK
- Expanded Block with new functions for new data stored in block
- Expanded debug_dump to also print out IN/OUT