Code generator use result from register assignment
The tests currently fail, the generated assembly is currently incorrect
as register a gets overridden.

To allow the code generator to use results from register assignment,
register assignment stores chosen register in Symbol instead of IGNode.
As some symbols reside on the stack, GRegister was expanded to include
the stack - hence its rename. How the stack size is calculated was
changed as not all variables are on the stack (the old method was to
could the size of every variable).
- Rename GRegister to Location
- Move reg_stack to Location
- Register assignment will not assign a register to those preassigned a
  location (e.g., function parameters)
- Expanded registers supported by reg_get (As register allocation uses
  all registers, need to support all of them)
- Changed symtab_add to use location none by default, previously
  defaulted to stack. Done as register assignment chooses location
  (register or stack)
- Added function to calculate stack size