Simplify register allocation/assignment algorithm
No need to check neighbors to determine which to spill, it is always the
current node as it has a lower spill cost since it is later in
occurrence (as the list is iterated from highest spill cost).

No need to check after completion to see if there is an opening for
spilled variables, as variables assigned registers never give the
registers away, and unassigned neighbors either get spilled or gain
a register, offering no openings in registers.