Switch to macro expansion for instruction selector
The capabilities of tree rewrite are limited, it is dependent on the
arrangement of expressions in the source code, and is mostly limited to
a single expression - providing limited opportunities for optimization.

For example:
a = a + 1; // Only forms a tree with this
b = 2;
a -= 2;
whereas:
a = a + 1; // Can form a tree with both these lines
a -= 2;
b = 2;

Thus the new plan is to implement a more advanced instruction selector
in the future capable of seeing and applying patterns across the entire
function. However for the initial implementation, a macro expander is
used to keep it simple while the other components are being implemented.
- Updated asmgen.md to reflect new plan for instruction selection