Add storage of pseudo-assembly in blocks
The macro replacement which requests a new symbol be implemented has yet
been implemented.

- Rename Statement to ILStatement
- Add struct PasmStatement
- Add debug printing of pseudo-assembly in blocks

Changed getters for InsSelMacroReplace to be by index of operand instead
of one getter for each operand, as it is easier to operate on all the
operands. As a result, the instruction selection macro operand type
"None" is no longer necessary as the macro now stores the number of
operands.