Work on assembly generator code generator
The new code generator will replace the INSTRUCTION_CG functions as it
is easier to implement new instructions and moves the role of optimizing
the code to a separate module of the assembly generator.

Changed instruction selection to store Register in PasmStatement instead
of Location, as assembly generation requires a register be provided.
However, certain registers can not be determined ahead of time, such as
"mov a, %some_symbol", the size of the symbol must be determined during
compilation. Thus, the instruction selection macros supports both
Register and Location, with Location being converted to a Register by
looking at the byte size of the other operands in the IL instruction.

Register/Location changes:
- Changed instruction selection macros to support Location and Register.
  Expansion of push and pop always uses QWORD registers, others such as
  mov specify Location and leave it to be computed at expansion.
- Changed PasmStatement to store Register instead of Location
- Added function to convert Register to Location (Used to calculate
  register preference for a PasmStatement operand which is now a
  Register)

Changes for code generator:
- Fix function returns being incorrectly implemented. Assembly was
  missing leave and ret
- Added instruction selector macro with no operands, used to implement
  x86 leave and ret
- Code generator aligns instructions and assembly to make them easier to
  read