IL gen for expression integer implicit conversion
Each of the cg_ functions for generating expressions will check the
types of the operands and promote or convert them to a common type as
necessary.

To reduce duplicate code, a single function was created to generate IL
for productions involving an operator with 2 operands. The function
figures out what IL instruction to emit by treating the operator in
the parse tree as the IL instruction, for example:
additive-expression:
 - a
 - sub
 - additive-expression:
   - b
   - add
   - additive-expression:
     - ...
Translates into IL:
add t1, b, ...
sub result,a,b

- Added new IL instructions for widening and narrowing of types
- Removed functions introduced in last commit which were planned to be
  used for implicit conversion, but were determined unnecessary
- Changed type_promotion to take Type as it is more convenient to use.
  The old version returned TypeSpecifiers which had to be inserted into
  a Type by the caller
- Update incorrect comment regarding type in test conversion.c
- Add new tests to check implicit conversion on unary operators