Add asm gen for IL instruction mse
The parser issues a mse to convert between two different C types, but if
they are actually the same byte size it leads to an invalid operation
with x86 movzx. Thus, mse has two cases: one if the operands are
different sizes, in which movzx is issued, and another where the
operands are actually the same size and a mov is issued.

To support the mse behaviour, the instruction selector constraint syntax
had to be extended.
- Constrain by the byte size of the symbol/constant
- Multiple constraints separated by spaces, only one of which needs to
  match
For example the constraint for mse: s2s1 s4s1 s8s1 s4s2 s8s2 s8s4

IL instruction mse could be merged with mov as its behaviour can be
implemented by mov. However, I believe allowing mov to implicitly copy
between different type sizes could be a source of error.

- Add IL instruction mse
- Add x86 instruction movsx
- Mark symbol_bytes const so it can be used with const Symbol
- Rename conversion.c to conversion_expression.c to be more descriptive