Work on read from/write to pointer
The indirection operator applied to a pointer either writes or reads
from the pointer, i.e., *p = a; or a = *p;. To handle the two possible
cases and propagate the correct type information for conversions, the *
operator returns an unnamed symbol. The symbol remembers the pointer so
IL instructions for the read/write to the pointer can be generated.
Lvalue and non Lvalue was implemented to decide on how to generate IL
instructions when referring to the unnamed symbol, if referring as a
Lvalue, it writes into the pointer - while if reading from the pointer,
it is first converted into a non Lvalue, which generates code to
dereference the pointer into a named temporary.

To allow symbols to represent being a Lvalue to the value at a pointer
and receive special handling when generating IL, symbols now have
classes. When generating IL the type of the symbol can be checked and
the appropriate instructions emitted.

- Add symbol classes
- Add value categories (Lvalue, non Lvalue)
- Allow symbols to be added to the symbol table with no name (to create
  the special Lvalue which represents a memory location when a pointer
  is dereferenced and can be written to/read from)
- Update IL instructions for memory address to allow indexing (mti, mfi)