Add value category conversion in operators
The C standard 6.3.2.1 specifies conditions when a Lvalue has to be
converted to a non Lvalue. Getting the conversions correct means
expressions such as a + *p - c are correctly handled as *p is converted
to a non Lvalue by dereferencing the pointer for results to be used in
the expression.

To reduce the risk of errors from forgetting to convert to a non Lvalue,
functions were added for IL instructions which convert the operands and
generate the IL. Sometimes the operands are required by the C standard
to remain as a Lvalue (operators &, ++, --), in which case they output
IL directly by calling parser_output_il.

- Add test to check Lvalue conversion to non Lvalue in arithmetic
  expression
- Add functions which converts operands to non Lvalue and generates IL
  instruction