Add IL generation for array subscripting
In order to store the array index, the special symbol which represents a
pointer location was extended to also hold the symbol representing the
index to reach the location.

Added ability to access multi-dimensional arrays by correctly handling
value categories for the indirection operator. On dereference, the
pointer Lvalue is converted to a non Lvalue to get the value of its
location so its location can be written to/read from.

- Fix incorrect counting of more than 1 pointers types
- Added tests for array subscripting