Add parsing of array declaration
It was realized that implementing arrays would require changes to the
type system as the arrays can be of a fixed or variable length, which
must be indicated the assembly generator via some mechanism. Currently,
the priority is on the implementation of pointers, thus only parsing is
implemented for now.

- Add tests for constant and variable length allocated arrays