Add asm gen for taking address of locals
To take the address of a variable, it must reside in memory. Thus a
precolor stage was added to the register allocator which will
automatically spill variables that have their address taken. However,
the spill code generator stage sees that the variable was spilled and
reloads it back into a register! To solve this, a table was added which
indicates which operands of a x86 instruction can be in memory, thus no
spill code needs to be generated. For optimize1.c this resulted in 818
fewer instructions from 2000 by not emitting spill code where possible.

- Add table of addressing modes for x86 instruction operands
- Spill code is not generated if there exists an addressing mode for
  accessing memory
- Add support for x86 instruction lea
- Add precolor stage to register allocator as lea requires operands in
  memory