Add ability to write to pointers
The IL instruction mti was implemented, which allowed pointers to be
written to. Offsets are currently NOT implemented, meaning it always
writes to the destination of the pointer.

To implement the instruction, a new attribute had to be added to the
instruction selection macros - dereference the symbol / register. To do
this, a flag system was added for the macros, where each flag is bitwise
or in. Doing so avoids having to add a macro parameter for each
attribute we need to add to a macro in the future, instead a single flag
parameter is taken.

- Add asm_gen support for IL instruction mti
- Add flag system for instruction selection macros, to change how the
  asm instructions are generated
- PasmStatement stores flag instead of size override
- Updated documentation on pseudo-assembly to allow memory addressing