Work on assembly generation for static arrays
The parser emits an array type in the IL to inform the assembly
generator of the array. The assembly generator will utilize instruction
selection macros specifically for arrays to generate the
pseudo-assembly, precolor the array to the stack, then calculate the
size directives in the final assembly using the size of the array's
elements.

Currently, the arrays are limited to a single dimension to keep the
initial implementation simple.

- Arrays are precolored to the stack
- Fix not calling type_construct() when creating a new Type, which led
  to uninitialized values.
- Update assembly generator debug print functions to print [] for
  pseudo-assembly memory access and symbol table arrays
- Add constraint character 'a' to instruction selection macros to
  constrain to array.
- Updated documentation on how arrays are implemented in the assembly
  generator

- Add functions for converting strings to integers
- Updated Type functions to support arrays
- Changed type_dec_pointer to behave as decreasing a level of
  indirection to be compatible with arrays. E.g., int[] -> int, int*[]
  -> int*

- Updated parse cg_ functions to handle arrays by storing a '(' or '['
  token during parsing to indicate whether it is an array of parameter
  list
- Updated cg_extract_symbol to also extract array type