Work on ability to index 1D arrays
The parser generates the expression to index the pointer in bytes, and
the assembly generator emits the appropriate instructions. Having the
parser calculate the index allows for the usage of more complex types in
the future (e.g., int a[100][20]; a[2]; each increment has a size of 100
ints), and simplifies instruction selection by avoiding having to
perform scaling to convert to a byte offset.

The assembly generator emits "mov [rbp+offset+index], some_value" to
access the stack, where [rbp+offset] is the location of the array and
index is a register holding the index. As of right now it does not
handle the cases where the index is a constant and when the index is
spilled.

- Updated intermediate language such that indices are in bytes.
- Add ability to store 2 symbols per pseudo-assembly statement to allow
  operand to be [base + scale].
- Add new addressing mode system for instruction selection macros, mode
  and parameters are packed together in one integer. Doing so allows for
  multiple parameters, previously only one was supported.
- Add new instruction selection addressing mode: OFFSET() which accesses
  the parameter 1 at an offset of parameter 2.
- Updated parser, when indexing array it emits IL instructions to
  convert index to bytes first, then performs the indexing.