WIP isel using custom instruction from x86 ins
Doing so simplifies the pseudo-assembly which has to be processed later
on, as the custom instructions contain no complex addressing modes -
only addressing a register, symbol, or dereferencing. (perhaps we should
separate out instructions with a dereference flag into a separate
instruction).

Previously for each pasm instruction, we have to check whether it is a
register or a symbol for each operand, then as well check for memory
accesses, multiple symbols in one operand, then for the dereference flag
- which makes working on pasm instructions a massive pain and very easy
to get wrong.

With the new pseudo-assembly instructions which are mapped to assembly,
e.g., mov_ss, mov_rs-> mov, operands have only one symbol each, can only
be a symbol or register, or the dereference flag. The complex addressing
modes are still possible, but they have been separated out to different
instruction names - e.g., mov_ss, mov_so for "mov sym, sym" and
"mov sym, sym" offset sym respectively.