Reduce spilling for assignment expression
Instructions for assignment expressions used to be generated by
generating the operand first, then instructions for the assignments.
This is however unoptimal for a long assignment expression such as the
following:

a1 = a2 = a3 = ... = a(n-1) = an
operand-1-instructions
operand-2-instructions
operand-3-instructions
 ...
operand-(n-1)-instructions
operand-n-instructions
assignment-n-to-(n-1)-instructions
assignment-(n-1)-to-(n-2)-instructions
 ...
assignment-2-to-1-instructions

The above results in a large number of live temporaries which must be
kept around (from operand 1 all the way until assignment 1) and
increases the chance of variables getting spilled. The new order is as
follows to reduce the duration of variables' lifetimes:

operand-n-instructions
operand-(n-1)-instructions
assignment-n-to-(n-1)-instructions
operand-(n-2)-instructions
assignment-(n-1)-to-(n-2)-instructions
 ...
operand-2-instructions
assignment-3-to-2-instructions
operand-1-instructions
assignment-2-to-1-instructions

For the test register_alloc_memory_addressing, the change resulted in a
reduction of x86 instructions from 537 to 326