Add compilation of multiple functions
The parser was extended to parse function calls and multiple functions.
Some changes had to be made to correctly handle symbols across
functions, i.e., symbols for functions are kept in global scope, types
for functions, and removing the f@ prefix prepended to the label at the
start of functions. The ability to call functions is still in progress.

- Add ability to parse argument-expression-list (part of function call)
- Update IL instruction call to have return value in arg0 (similar to
  other instructions)
- Add new type category: functions
- Function names in assembly are no longer prefixed f@ so they can be
  called by programs from other compilers
- Assembly generator remembers function names when the symbol table is
  cleared at the end of a function. This is done by only clearing a
  portion of the symbol table - the portion which corresponds to the
  symbols within the function.