Remove unnecessary push in division + fix division
The current implementation of division always pushes off rax and rdx,
the push instruction uses the first operand, which liveness analysis
propagates. If the push is unnecessary, it propagates the operand all
the way to the beginning of the function (since it is never defined).
This causes the call instruction to see that it needs to keep the
operand live, and thus results in extra push + pop for the function
call.

Now, division pushes only if it sees a live value in the A or D
register. The benefit of less live values unnecessarily propagated also
leads to less nodes in the interference graph, leading to faster compile
times.

It was also discovered that division was incorrectly implemented and has
several issues. Integer division needs to sign extend the A register to
D, and required a different instruction for integer / unsigned division.
Always pushing and popping rax and rdx may overwrite live values:
push            rax
push            rdx
xor             edx, edx
mov             ebx, 255
idiv            ebx       <-- Remainder is stored in edx
pop             rdx       <-- Remainder incorrectly overwritten
pop             rax
mov             eax, edx  <-- Reading incorrect value for remainder

Incorrect liveness calculation for dividend, as it does not see it being
used in idiv:
mov             eax, some-val
mov             eax, 255      <-- Incorrectly overwrites A
idiv            eax

- Add x86 instructions cwd, cdq, cqo, div, xchg
- Changed maximum symbols which can be returned by pasmstat_def to
  MAX_ASMINS_REG (xchg defs 2)
- Add test for signed and unsigned division
- Avoid having to manually add each special pasm instruction handler by
  having macro expand out switch cases