Add container which does not invalidate pointers
The current dynamic array invalidates pointers when it resizes, which is
a source of subtle bugs which only occurs under certain circumstances -
when the array resizes. To solve the problem, in places where the
current the dynamic array needs to be used while its elements are
accessed, a new container (hvec) which does not invalidate pointers on
resize is used instead.

The new container stores the objects on the heap, and stores a pointer
to the object. On access, a pointer to the object itself is returned,
not to the vec. This allows the container to change without invalidating
existing pointers to the objects in use.

The vec switched over to hvec are the symbol table and the block
PasmStatement. As a result, extra code written to reload symbols /
PasmStatement to guard against resizes were removed.