Avoid attaching unnecessary nodes to tree
Nodes are only attached to the tree if they are valid. Previously it
always attaches a node, leaving it in a invalid state if the production
failed to match.

To do so, tree nodes are heap allocated to allow manipulation in each
parse function before attaching it to the tree.