Add asmgen to be compiled by make
To support the original set of compiler warnings, build.sh passes them
to make. Thus build.sh should be called to make the program, not make
itself.

The source files were adjusted slightly to compile and resolve any
warnings. The test files compiles with a less stringent set of warnings.