Store Symtab Symbols until Symtab destruction
Previously, symbols were only stored until the end of scope. This is a
problem as the IL2 generator cannot access the symbols anymore. Now, the
symbols are stored until the symbol table itself is destructed.

As a result of how symbols are now stored - in one memory location
without moving - they no longer need to be referenced via an id, their
pointer can be directly returned