Setup Tree traversal for generating IL2
The traversal motion required for generating IL2 into the CFG is
complete. Now, the old code for actually generating the IL2 must be
brought over.

Changed parse declaration to also remove the assignment expression node
if it is unnecessary. Otherwise, it has to determine if the assignment
expression node is valid when generating IL2.